Precompile templates for a speed improvement
Previously, for every request to /, the app would have to read
site/index.html.eex, parse it, and only then actually run it. Now, we
compile all templates at startup to eliminate the first two steps of
that process. This makes the time to process a given request shorter
and should allow the app to do less work and less IO under load.

This is a good design considering we might make pages using templates
which are actually dynamic, but the ones we actually have so far are
the same every time, so we could further optimize by just caching the
final text output.